This component is called from footer as well so move it
